+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/demandandcapacitynotification/logic/service/ReportedDemandAndCapacityNotificationService.java
@@ -52,7 +52,8 @@ public class ReportedDemandAndCapacityNotificationService extends DemandAndCapac
                 (
                     (
                         notification.getStatus() == StatusEnumeration.OPEN &&
-                        notification.getResolvingMeasureDescription() == null
+                        notification.getResolvingMeasureDescription() == null &&
+                        notification.getText() != null && !notification.getText().isBlank()
                      ) ||
                     (
                         notification.getStatus() == StatusEnumeration.RESOLVED &&